+++ spec/support/puts_debuggerer_invoker.rb
@@ -55,4 +55,8 @@ module PutsDebuggererInvoker
   def self.vararg_array_with_options(options)
     pd 'hello', 3, true, options
   end
+  # intentional empty line
+  def self.object_with_return_option(object, return_option = nil)
+    pd object, {}.tap {|h| h.merge!(return: return_option) unless return_option.nil?}
+  end
 end